wallet-api: Extend Ledger.Validation
* Merge Language.PlutusTx.Validation (plutus-use-cases) into
  Ledger.Validation (wallet-api)
