+++ plutus-use-cases/src/Language/PlutusTx/Coordination/Contracts/Vesting.hs
@@ -23,7 +23,6 @@ import           GHC.Generics                 (Generic)
 import           Ledger.Validation            (Height (..), PendingTx (..), PendingTxOut (..), PendingTxOutType (..),
                                               ValidatorHash)
 import qualified Language.PlutusTx            as PlutusTx
-import qualified Language.PlutusTx.Validation as PlutusTx
 import           Ledger                       (DataScript (..), PubKey (..), TxOutRef', ValidatorScript (..), Value (..), scriptTxIn, scriptTxOut)
 import qualified Ledger                       as Ledger
 import qualified Ledger.Validation            as Validation
@@ -111,10 +110,10 @@ validatorScript v = ValidatorScript val where
         let
 
             eqBs :: ValidatorHash -> ValidatorHash -> Bool
-            eqBs = $$(PlutusTx.eqValidator)
+            eqBs = $$(Validation.eqValidator)
 
             eqPk :: PubKey -> PubKey -> Bool
-            eqPk = $$(PlutusTx.eqPubKey)
+            eqPk = $$(Validation.eqPubKey)
 
             infixr 3 &&
             (&&) :: Bool -> Bool -> Bool
